[kernel emul] fix a bug in which emulated kernel enqueue still use uint32_t pointers for arguments

+++ c10/hammerblade/emul/bsg_manycore_cuda.cpp
@@ -307,7 +307,7 @@ void reset_runtime() {
                                        hb_mc_dimension_t tg_dim,
                                        const char *name,
                                        const uint32_t argc,
-                                       uint32_t *argv) {
+                                       uint64_t *argv) {
           EMUL_WARNING();
           if (!device_busy || !binary_loaded) {
             return HB_MC_UNINITIALIZED;
